+++ speech.py
@@ -22,17 +22,27 @@ class Speech:
         self.character_end_times = speech['alignment']['character_end_times_seconds']
 
     def GetWordDurations(self):
-        words = []
-        start = 0
-        for i in range(len(self.characters) + 1):
-            if (i == len(self.characters) or self.characters[i] == ' '):
-                words.append({
-                    'word': ''.join(self.characters[start:i]),
-                    'start': self.character_start_times[start],
-                    'end': self.character_end_times[i-1]
-                })
-                start = i
-        return words
+         '''
+         Returns an array of dicts containing the word, start time, and end 
+         time (seconds) for each word in the speech
+
+         Args:
+            None
+
+        Returns:
+            [dict]: Array of dicts with keys: text, start, and end
+         '''
+         words = []
+         start = 0
+         for i in range(len(self.characters) + 1):
+             if (i == len(self.characters) or self.characters[i] == ' '):
+                 words.append({
+                     'text': ''.join(self.characters[start:i]),
+                     'start': self.character_start_times[start],
+                     'end': self.character_end_times[i-1]
+                 })
+                 start = i
+         return words
     
     def CompressToMp3(self):
         '''